+++ app/src/main/java/com/ar/arstoken/data/db/SaleEntity.kt
@@ -16,9 +16,9 @@ data class SaleEntity(
 
     val saleType: String,
 
-    val totalAmount: Int,
-    val paidAmount: Int,
-    val dueAmount: Int,
+    val totalAmount: Double,
+    val paidAmount: Double,
+    val dueAmount: Double,
 
     val synced: Boolean = false,   // 🔑 cloud-ready
     val updatedAt: Long = 0
