+++ app/src/main/java/com/ar/arstoken/model/ReportSummary.kt
@@ -1,8 +1,8 @@
 package com.ar.arstoken.model
 
 data class ReportSummary(
-    val totalSales: Int,
-    val totalCash: Int,
-    val totalCredit: Int,
-    val totalDue: Int
+    val totalSales: Double,
+    val totalCash: Double,
+    val totalCredit: Double,
+    val totalDue: Double
 )
